Remove most custom styling for the launch button
We can put it back later with a larger redesign. The grey just
made the buttons look disabled with the rest of the theme.

This also overrode some of the base styles from bootstrap (such
as form-control) unrelated to what it was styling. Reverted that
too.

Use btn-warning (orange), since Freeman's initial objection was
to 'bootstrap blue'