Add previous and next link to the bottom content in each page.
This is in addition to the content in the sidebar in order to be more
discoverable by users.

I'm also personally a fan of having it on top of each page as well; but
I don't think it fits well with _this_ theme, And the link in the
sidebar are sufficient.