Add top Previous/Next on each page.
As requested by willinc, to see what it looks like (and if we like it).
Text is literally `Previous` and `Next` not to visually change between
page and stay discreet.

I think it is useful, (I like that), but I tink it is only minimally
useful so would be ok if it's not merged.