Prototype: Federation settings.
This is a prototype (have to sanitize input) on how federation would be
enabled.

By some mechanism a user is redirected to https://mybinder.org/register/https%3A%2F%2Fbinder.example.com
(the last part being https://binder.example.com), binder thus set a
cookie that lists binder.example.com as a potential redirect target, and
redirect user to the /settings/ page.

Here a user have the choice to select their default binder. We set
by default _always_ to mybinder.org, unless user action. The settings
page can set the _default_ binder (in a cookie).

Now we "just" need to setup the /v2/ uri to lookup this default and
redirect to it. And you get mybinder.org to be a "simple" trampoline to
other binders.