support handcrafted links to non-github binders
The index.html form still support only GitHub repositories, but you
should now be able to handcraft links to binders with the gitlab or
git providers.

* binderhub/main.py (ParameterizedMainHandler): Pass the
provider_prefix to the loading.html template.
* binderhub/repoproviders.py (GitRepoProvider): Set unresolved_ref
as a copy of resolved_ref, because ParameterizedMainHandler
reads the former.
* binderhub/templates/index.html, binderhub/templates/loading.html:
Add a hidden field storing the provider_prefix.
* binderhub/static/js/index.js: Adjust to take the provider_prefix
into account.