remove specified version
pinning a specific tag gives the impression that it's a good choice.

We can pin a specific version later once we make releases,
but while we are in "always master" mode, we shouldn't give the impression that a specific tag in the past is the right version.

Instead, link to the helm chart recent releases list.