split per-repo and timing metrics
each unique value for each label has a cost in prometheus

The timing metrics right now have lots of axes (status, retries, bucket, repo),
resulting in a bit of an explosion.

It is especially an impediment when attempting to look at summary statistics
where the repo is not relevant
because prometheus must sum over many per-repo vectors.

This removes the repo label from timing metrics and adds counter metrics for the repos,
allowing us to hugely reduce the cost of overview metrics,
such as build time and success.
Additionally, it reduces the resolution of our timing buckets in the histograms.

Changes:

- reduce buckets for histograms, with separate buckets for builds and launches
- don't resolve builds that take less than a minute or launches that take more than 10
- remove provider, repo metrics from build/launch time metrics
- add (launch|build)_count Counter metrics, with status, provider, repo

The only downside I can think of:

It was technically possible, but we never used prometheus to identify
which repos were the slowest.
This will no longer be possible.