Remove travis_wait from tests
travis_wait should only help when a test produces no output for ten minutes.
This shouldn't happen for us when we are running tests with `-vsx`

The downside of using travis_wait is that it suppresses all output until the tests finish,
so if anything causes a hang or other early failure,
we won't be able to see it until travis_wait times out.