Drop restriction on num of imgprefix path elements
Accept any number of elements in the image prefix passed to
get_image_manifest. Add support for gitlab registries which are
per project and can have things like: <group>/<project>/<image>:<tag>.

Or even deeper structures if subgroups are used.

It implies only the real image part is being set in image_prefix, not
the host (which i assume is the case). So something like
rochaporto/myimage:mytag is ok, but not docker.io/rochaporto/myimage:mytag.