Fix CI tests - Running != Ready
There was a test failure because the hub became running but later
crashed, these kinds of crashes can happen but the hub will not become
ready until the risk of it happening has gone away. So, if we wait for
the hub to be Ready as compared to only become Running, we are good.

For this to make a difference, Running must be distinguished from Ready,
which it will become in a Z2JH release where we introduced a Readiness
probe. What version of z2jh are we using?