Bump chartpress to 0.5.0
Chartpress 0.5.0 contains a bugfix that makes a chart's images' context
paths be considered when determening the version of the chart as well as
the chart folder. This was not done before and caused issues, that made
us require https://github.com/jupyterhub/binderhub/pull/1014 which can
now be reverted in this commit.