CI: remove broken code in modern minikube
Disabling the dashboard may make sense in production environments, but
is pointless in CI. So since the command breaks in modern version of
minikube, I removed it.

Also, the environment variable will be reset when using sudo with
minikube, so due to that, it doesn't make sense to set in .travis.yaml
but instead only set during startup of minikube.