ci: bugfix binderhub image not updating on change
Chartpress 1.0.0 contained a regression, and our publishing pipeline was
using chartpress 1.0.0 instead of 0.7.0, so it was affected. This made
the image version used by binderhub since chartpress 1.0.0 was released
in 21 Nov 2020 be the same version as the image-cleaner image.

The only PR that was influenced by this mishap was a PR merged ~2 hours.
With this commit, it should not happen again.