Retrieve baseUrl & build token just before they are needed
The bundle is included in the <head>, so code in top level
might be executed before the DOM is loaded. I think this was
happening, causing errors of form `$(...).data() is undefined`.

Since they are used in only one place, we instead retrieve just
before they are used. The `indexMain` entrypoint is called from
the footer, so this code will execute after the DOM has
loaded.

We can also pass parameter name to `.data()`, which will return
undefined if it doesn't exist in the element. Easier than checking
if the output of `.data()` is undefined.