chart: add an optional podman service pod
This is not a kubernetes service but podman running as a service which
allows for a client/server setup as we can currently do with docker.

This service can be used in place of the one defined by dind so podman
can be used as a build engine while keeping the cache of images
available. This is a privileged pod in the same fashion as the dind one
but it runs as the podman user. The builder pods can then connect to the
service to do the actual work.

The current Build class makes some legitimate assumptions about docker
use which is not an issue except for some naming mismatch that could be
a bit misleading when setting up podman as replacement. It also does
some affinity setup which we have to currently lie to in order to create
the builder pods on the same machine as the service pod.

Finally, there are two environment variable that must be set in order
for the client to be properly configured. The first one is
CONTAINER_HOST which will trigger the "remote" mode of podman and the
second is REGISTRY_AUTH_FILE to ensure podman grabs the proper Docker
credential file. Currently these requires to re-implement the submit
method of the builder class as there is no way to pass them as
configuration options.