feat(build): add configurable environment variables
Up until now the build pod did not need any custom environment variable.
However, in order for podman to work as we would like, this is a simple
and effective way. The other option would be to modify the podman engine
and add support for additional arguments there that could pass similar
values.

For example the equivalent of
  CONTAINER_HOST="unix:///var/run/docker.sock"
as arguments would be:
  --remote --url unix:///var/run/docker.sock

This would require a bit of redesign to the repo2podman engine in order
to pass these parameters everywhere a command is called. The benefit of
the environment variables is that it allows to change the behaviour
without code change.