refactor(helm-chart): Simplify image-cleaner handling
This patch refactors the creation and handling of the image cleaner daemonset.

The current behaviour makes it so that if one doesn't disable the host cleaner
explicitly when using dind (or pink), they would end up with two cleaners
instead of one.

The new behaviour avoids this issue by ensuring that only the container matching
the selected imageBuilderType is created. Same goes for the volumes handling.

In order to have consistent naming, the "local" imageBuilderType enumeration
value is replaced with "host".

It also deprecates the use of the imageCleaner.host.enabled parameter as you
either want image cleanup or not.