Remove redundant kubernetes version pin
It's already specified in the requirements.txt in the
root of the repo, so everyone developing binderhub locally
has probably been using a much newer version anyway.

I've left the `jupyterhub` pin in as it might need to be
co-ordinated with the version of z2jh the chart uses.

Fixes https://github.com/jupyterhub/binderhub/issues/1494